move devices to own endpoint
