+++ src/main.rs
@@ -100,10 +100,12 @@ async fn main() -> std::io::Result<()> {
                 )
                 .service(web::scope("/device")
                     .service(routes::device::get_device)
-                    .service(routes::device::get_devices)
                     .service(routes::device::create_device)
                     .service(routes::device::update_device)
                     .service(routes::device::delete_device)
+                )
+                .service(web::scope("/devices")
+                    .service(routes::devices::get_devices)
                 )
             )
     })
