+++ src/routes/mod.rs
@@ -1,2 +1,3 @@
 pub mod wake;
 pub mod device;
+pub mod devices;
